requirements.txt
Pillow>=10.0.0,<11.0.0


